requirements.txt
pycryptodome>=3.23.0
opencv-python>=4.0.0
numpy>=1.20.0
